refactor: extract chart expand state and esc key binding into useExpandedChart hook
